requirements.txt
flask
gluoncv
matplotlib
mxnet
requests
Pillow